requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coverage==4.4.1
dj-database-url==0.4.2
Django==2.2
django-environ==0.4.4
gunicorn==19.7.1
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pytz==2017.2
requests==2.18.4
selenium==3.5.0
six==1.10.0
urllib3==1.22
whitenoise==3.3.0
wrapt
djangorestframework
Pillow